cloud_run/requirements.txt
fastapi
uvicorn[standard]
pandas
google-cloud-storage
google-cloud-bigquery
pyarrow
pandas-gbq